fix error by Hung

+++ src/main/java/com/quafresh/web/aquafreshweb/repositories/DiscountRepository.java
@@ -4,5 +4,5 @@ import com.quafresh.web.aquafreshweb.entity.Discount;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DiscountRepository extends JpaRepository<Discount, Long> {
+public interface DiscountRepository extends JpaRepository<Discount, Integer> {
 }
